Specify output I2C port for NeoPixels display
Formerly, NeoPixel I2C displays were limited to the kOnboard I2C port. This
patch allows the calling class to specify which I2C port to communicate on.
This is especially necessary since the kOnboard I2C port has developed a
tendency to hard-lock the RoboRio in the wrong circumstances, necessitating
a switch to the kMXP I2C port.