Add SynchronousAnimator
The SynchronousAnimator is an Animator that relies on periodic refreshes
from animatePeriodic() instead of running the animator logic in a separate
thread. The goal is to avoid multithreading, which can be finnicky on the
robot.

This change also renames Animator to AsyncAnimator, and creates a new common
interface Animator for both AsyncAnimator and SynchronousAnimator.