AnimationCompositor rewrite
Previously Views were constructed separately and associated with a given
AnimationCompositor using a String key.

This patch changes this interface such that Views are constructed by the
AnimationCompositor at the client's request.

This simplifies the workflow for creating context-sensitive animations,
since now a Subsystem can just grab a view at startup from the
AnimationCompositor and use the .show() and .hide() methods on the view to
show/hide the associated animation.