Try to fix SPI endianness issues
It seems the rio is configured to send 16-bit words at a time.
However, the arduino is expecting 8-bit bytes at a time. This creates
an issue if the rio sends its words in little-endian, since the arduino
will see it as the bytes being sent in reverse order.
This patch (hopefully) fixes this issue.